+++ src/nba_impact/data/__init__.py
@@ -1,1 +1,0 @@
-"""The data layer of nba-impact."""