+++ src/nba_impact/rapm/__init__.py
@@ -1,1 +1,0 @@
-"""The rapm layer of nba-impact."""