+++ src/nba_impact/reliability/__init__.py
@@ -1,1 +1,0 @@
-"""The reliability layer of nba-impact."""