+++ src/pippen/model/__init__.py
@@ -1,0 +1,1 @@
+"""The model layer of pippen."""